Comment out Numba
It looks like jit, njit and prange are imported from Numba in ATS.py file and they are not used in the code and this is the only place numba is called. So the change is to remove it from requirements.txt file and comment it out in ATS.py file.